Extend parser using arguments provided by the sources
Instead of recursively iterating through the models' API to add the cli
arguments, take them from the sources. Each source will provide only the
arguments that it supports, so the cli will reflect more accurately the
capabilities according to the configuration. If an argument is found in
multiple sources, then its level is set to the maximum of the values
found. This allows us to have an argument defined at multiple sources
and place it at the right level.